Can view CartPole while learning
Simply adding the command env.render() in the episode loop allows us to watch the policy executing.

This basically eliminates the need for dqn_agent, so I deleted it.

In fact, this code runs fine in isolation ... it doesn't need the server to be launched. It sort of begs the question of whether this was the right repo to start with ... the tidy RL repo seems to have more or what we want.

Still, now that we have a working DQN, we can expand on things.

#1