Update requirements.txt #13
The requirements file has been updated according to the domains we are testing our algorithms in. Note that because of the nature of the gym-retro and atari-py packages, their libraries (gym, retro, and atari) must be installed beforehand. Additionally, gym-retro==0.7.0 installs gym==0.12.1 and its supplemental packages (pyglet; scipy; numpy, which was already taken care of in the preceding line; six; and future)