Solved CUDA device issue
When @nazaruka copied the PPO code he was working on a CPU only machine, so he hard coded a to("cpu") command which is why I was having CUDA problems. Although I had fixed it in my own way, I found a more elegant solution that looks more like the original code, and have reverted things a bit.