Watch learning and eval
The code is currently set up to display both the learning and the evaluation, and the results are very different. During learning, epsilon causes Sonic to jump around a lot, and sometimes get far as a result through luck.

Eval behavior is more consistent, but as a result often charges blindly forward and dies or gets stuck. However, Sonic doesn't always die on that first robot, so that is some progress.