Fixed mutation and generation of next generation
The NSGA-II code we had was assuming that genomes were single numbers, not arrays. So, generated children were single numbers. This has been fixed.

We were also generating twice as many children as needed, which was an error.

I made some comments pointing out magic numbers that should probably be command line parameters in the future.