This mutation is much faster
Instead of looping through the array, I simply replace the solution with a normally distributed vector where the centers are the values in the original solution. However, this may be too random.

Simply reducing the scale from 3 to 1 could help a lot, but there is one other issue: Now all indices are mutated ... need a way to only mutate some but not all. Perhaps this could be done by making the scale be a vector with 0s in it.