PPO has no epsilon
I've just clarified an error in my understanding of PPO. In typical RL algorithms, epsilon indicates the percentage of the time that the agent takes a completely random action. However, PPO doesn't do that. Rather, it starts with a policy that outputs action preferences, but it chooses actions from that policy stochastically rather than deterministically during learning. This is the only source of randomness in action selection, and is those the source of all exploration.

I have questions as to whether this is working properly still ... though PPO might simply be converging to a suboptimal policy. The fact that we are not doing Joint PPO might also make premature convergence more likely.

I will close issue #29 because of this.
@nazaruka