Init weights from network
The CNNs are randomly initialized, but the way their weights are initialized is more sophisticated than how we are creating random genomes. These initial weights lead to obviously better policies ... the agents explore more, making it possible to actually learn which actions are good and which are bad. When we use the completely random genomes, the agents just get stuck in very suboptimal behaviors very quickly. #27