Removed default params
For all batch files, if there was a setting equal to the default setting in arguments, I removed it (redundant). This led me to some discrepancies, which I also fixed.

Currently, I've left in --use-proper-time-limits because the default is False, but I'm not sure I actually want to use it. I'm a bit uncertain about linear decay as well, but for now it defaults to True.